backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
sqlmodel==0.0.14
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
